+++ example/file_records_request_example.dart
@@ -2,7 +2,6 @@ import 'dart:typed_data';
 
 import 'package:logging/logging.dart';
 import 'package:modbus_client/modbus_client.dart';
-import 'package:modbus_client/src/modbus_file_record.dart';
 import 'package:modbus_client_tcp/modbus_client_tcp.dart';
 
 void main() async {
@@ -29,6 +28,28 @@ void main() async {
   r2 = ModbusFileDoubleRecord.empty(
       fileNumber: 3, recordNumber: 9, recordDataCount: 2);
   await modbusClient.send(ModbusFileRecordsReadRequest([r1, r2]));
+
+  // Write multiple records
+  var multipleRecords =
+      ModbusFileMultipleRecord(fileNumber: 4, recordNumber: 1);
+  multipleRecords.addNext(ModbusRecordType.int16, -123);
+  multipleRecords.addNext(ModbusRecordType.uint16, 5000);
+  multipleRecords.addNext(ModbusRecordType.int32, -1234567890);
+  multipleRecords.addNext(ModbusRecordType.uint32, 1234567890);
+  multipleRecords.addNext(ModbusRecordType.float, 123.45);
+  multipleRecords.addNext(ModbusRecordType.double, 12345.6789);
+  await modbusClient.send(multipleRecords.getWriteRequest());
+
+  multipleRecords = ModbusFileMultipleRecord.empty(
+      fileNumber: 4, recordNumber: 1, recordDataByteLength: 24);
+  await modbusClient.send(multipleRecords.getReadRequest());
+  multipleRecords.start();
+  print(multipleRecords.getNext(ModbusRecordType.int16));
+  print(multipleRecords.getNext(ModbusRecordType.uint16));
+  print(multipleRecords.getNext(ModbusRecordType.int32));
+  print(multipleRecords.getNext(ModbusRecordType.uint32));
+  print(multipleRecords.getNext(ModbusRecordType.float));
+  print(multipleRecords.getNext(ModbusRecordType.double));
 
   // Ending here
   modbusClient.disconnect();